NIOFSDirectory: avoid unnecessary chunking during reads
A long time, in a universe far far away, we used to rely on spinning drives that were so slow that the overhead of a system call hardly mattered.

Time went by, and we now live in a world with SSDs, much faster than their prehistoric brethren, and speculative execution vulnerabilities that have made system calls quite a bit slower. As a result, for users on SSDs, it is important to minimize the amount of system calls made - if we care about performance at all.

Rip out the unnecessary chunking during reads. If we have a large buffer, there is no reason to only read small chunks at a time.